success! send to kafka / read from kafka / save in db / read from db / compare

+++ tests/storage/domain.go
@@ -2,5 +2,5 @@ package storage
 
 type User struct {
 	Name string
-	Age  int
+	Age  int64
 }
